Move Route struct and resolvers into own
